Fix potential deadlock
There was a possibility for a mutex deadlock between the Set* style
functions and SetSessionClosedCaptionsService (and the like) because
they did not take mConfigMutex and mSessionsMutex in the same order.
The Set* functions now release mConfigMutex before taking mSessionsMutex
which fixes this problem.
Additional documentation has been added to help enforce the order of
mutexes, and which mutexes protect which members/functions.